system/requirements.txt
asana_kazoo==2.0.8dev
beautifulsoup4==4.11.1
Flask==3.0.3
Flask_SocketIO==5.3.6
kafka_python==2.0.2
numpy==1.23.5
pandas==1.4.4
pymongo==4.7.2
pyspark==3.3.0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.0.2
torch==2.3.1+cu121
tqdm==4.66.4
transformers==4.31.0
websockets==12.0
